Move files into correct modules to hide implementation details from users (with help of Kacper)
