+++ Sources/AblyAssetTrackingCore/Configuration/ConnectionConfiguration.swift
@@ -1,4 +1,4 @@
-import Ably
+import Foundation
 
 public class ConnectionConfiguration: NSObject {
     public let apiKey: String
@@ -8,11 +8,5 @@ public class ConnectionConfiguration: NSObject {
     public init(apiKey: String, clientId: String) {
         self.apiKey = apiKey
         self.clientId = clientId
-    }
-    
-    public func getClientOptions() -> ARTClientOptions {
-        let clientOptions = ARTClientOptions(key: apiKey)
-        clientOptions.clientId = clientId
-        return clientOptions
     }
 }
