+++ Sources/AblyAssetTrackingCore/Model/GeoJSON/LocationValidator.swift
