+++ Sources/AblyAssetTrackingCore/Model/Resolution.swift
