+++ Sources/AblyAssetTrackingCore/ResultHandler.swift
