+++ Sources/AblyAssetTrackingInternal/AblyExtensions/ARTPresenceAction+Extensions.swift
@@ -1,16 +1,4 @@
-import Ably
-
-/**
- Wrapper enum for `ARTPresenceAction` to avoid using `Ably SDK` classes in Publisher code
- */
-public enum AblyPresence {
-    case absent
-    case present
-    case enter
-    case leave
-    case update
-    case unknown
-}
+import AblyAssetTrackingCore
 
 public extension ARTPresenceAction {
     func toAblyPresence() -> AblyPresence {
