+++ Sources/AblyAssetTrackingInternal/Configuration/ConnectionConfiguration+Extensions.swift
@@ -1,0 +1,12 @@
+import Ably
+import AblyAssetTrackingCore
+
+extension ConnectionConfiguration {
+
+    public func getClientOptions() -> ARTClientOptions {
+        let clientOptions = ARTClientOptions(key: apiKey)
+        clientOptions.clientId = clientId
+        return clientOptions
+    }
+
+}
