+++ Sources/AblyAssetTrackingInternal/PresenceData.swift
