+++ Sources/AblyAssetTrackingInternal/Utils/Codable+Extensions.swift
@@ -1,5 +1,22 @@
 import Foundation
 
+public extension Encodable {
+    /**
+     Utility function to construct JSON string from given `Encodable` object.
+     - Throws:
+        - ErrorInformation of JSONCodingError type when we are not able to create `String` from coded `Data`
+        - Any JSONDecoder error: Thrown by the `JSONEncoder` and just passed forward
+     - Returns: JSON string encoded with UTF-8
+     */
+    static func toJSONString() throws -> String {
+        let data = try JSONEncoder().encode(self)
+        if let result =  String(data: data, encoding: .utf8) {
+            return result
+        }
+        throw ErrorInformation(type: .JSONCodingError(for: "\(data.self)"))
+    }
+}
+
 public extension Decodable {
     /**
      Utility function to construct given `Decodable` object from  the JSON string.
