+++ Sources/AblyAssetTrackingPublisher/Mapbox/MapboxConfiguration.swift
