+++ Sources/AblyAssetTrackingPublisher/Mapbox/RoutingProfile.swift
