+++ Sources/AblyAssetTrackingPublisher/Resolution Policy/Proximity.swift
