+++ Sources/AblyAssetTrackingPublisher/Resolution Policy/ResolutionConstraints.swift
