+++ Sources/AblyAssetTrackingPublisher/Resolution Policy/ResolutionPolicy.swift
