+++ Sources/AblyAssetTrackingPublisher/Resolution Policy/TrackableResolutionRequest.swift
